#715: weigthed search

+++ factcast-store-pgsql/src/main/java/org/factcast/store/pgsql/registry/transformation/chains/TransformationChainMetaData.java
@@ -21,5 +21,5 @@ import lombok.Value;
 public class TransformationChainMetaData {
     String id;
 
-    int score;
+    long score;
 }
